Eliminate double JSON serialization for efficiency. (#9)
Eliminate double JSON serialization for efficiency.

Previously, field data would be marshalled within a call to `Event.Add`, in
order to check if it could actually be marshalled. Any fields with
values that can't be marshalled (channels, maps with non-string keys, etc.)
would be quietly discarded.

This is inefficient, since we have to marshal the event at transmission time
again anyways.

Instead, defer all marshalling to transmission time, but use custom marshalling
to ensure that any values that can't be marshalled are quietly dropped, instead
of causing an error.

This likely won't make much difference for most users, but does improve
efficiency for high-volume use cases (like the TCP agent).