Improve error responses for top-level HTTP errors (#16)
Improve error responses for top-level HTTP errors

Previously, when submitting event batches to the Honeycomb API, we'd always try
to unmarshal the server's response into a slice `[]*Response`.

This could lead to very confusing results in some cases. For example,
if you supplied a bad write key, the server would respond with
`{"error":"unknown Team key - check your credentials"}`. That can't be
unmarshalled into the type `[]*Response`, so instead the `Err` field on
responses would be populated with
```
json: cannot unmarshal object into Go value of type []libhoney.Response
```

To make that less cryptic, check the top-level HTTP response code before unmarshalling.
If that response code isn't 200, we set the status code and body from the
HTTP response on each individual `Response` that we put in the response queue.
That makes it a lot easier to figure out that events aren't being sent because
of, say, a write key misconfiguration.

Test plan: unit tests