add discard output implementation to make it easy to configure libhoney to be mute (#24)
While the default Writer implementation of the Output interface is good for development, and the Mock implementation good for when you want to test what libhoney is doing, the Discard interface is there to mute libhoney when you are testing other things (that might trigger sending events) and you want to send them to /dev/null.

This is often available in libraries as a mute option.