better debugging through a logger (#31)
To better enable people to debug their use of libhoney, this PR adds a Logger interface along with two default implementations - one to `stdout` and one to `/dev/null`. The user of libhoney is free to pass in their own logger, so long as it implements the `Printf(string, ...interface{})` method (which ,coincidently, the default `log` package implements).

The read json example is improved to highlight the `stdout` logger as well as actually succeed in going through the response queue due to adding a wait group. (previously it would exit before it could finish reading the responses.)
 
The logger is rather verbose, but I think that's ok for development. Here is some sample output:

```2018-09-21T16:46:57.437671909-07:00 - libhoney - DEBUG - adding event to transmission; queue length 0
2018-09-21T16:46:57.437678083-07:00 - libhoney - DEBUG - Send enqueued event with fields map[json_file_name:./example1.json json_line_number:5 read_json_log_version:0.0.1-example num_goroutines:3 timestamp problem:missing timestamp ingredient:eggs amount:3 step:5 comment:missing time field; this event will be sent with the current time.]; err=<nil>
2018-09-21T16:46:57.438657518-07:00 - libhoney - DEBUG - closing libhoney
2018-09-21T16:46:57.438663186-07:00 - libhoney - DEBUG - default transmission stopping
2018-09-21T16:46:57.995086199-07:00 - libhoney - DEBUG - got response code 202, error %!s(<nil>), and body
2018-09-21T16:45:59.056881115-07:00 - libhoney - DEBUG - got response code 401, error %!s(<nil>), and body {"error":"unknown API key - check your credentials"}
2018-09-21T16:45:59.056886224-07:00 - libhoney - DEBUG - got response code 401, error %!s(<nil>), and body {"error":"unknown API key - check your credentials"}
2018-09-21T16:46:57.995122157-07:00 - libhoney - DEBUG - got response code 202, error %!s(<nil>), and body
```