[transmission] reduce batch size if total payload will exceed 5MB
We default to a maximum batch size of 50 events. This usually works, but some users have large individual events, and can exceed the max json payload size for `/1/batch` of 5MB defined by the API. This causes the entire batch to fail. Instead of failing, we should make an effort to split up the batch and reenqueue events that don't fit into the first payload.