[transmission] Allow toggling of gzip compression on send
There are some use cases where CPU is scarce but network bandwidth is readily available, and turning of GZIP is useful. This adds an option to the default tx implementation to turn off GZIP. When initializing libhoney clients, a custom definition for TxDefaultClient can be passed as the Output to disable gzip.