Add a client object to libhoney for isolating traffic (#40)
This change adds a `Client` object to libhoney.

When sending traffic of different importance, it can be frustrating to have them run through the same transmission queue. It would be nice to be able to identify some traffic as important enough to block and slow down if sending backs up while allowing other (say logging) traffic to be dropped. With one transmission and a global `BlockOnSend` that's not possible. This change introduces a libhoney Client with its own queues and transmission. It does not share any state with the global client.