protect against the first event in a batch failing while others succeeed (#46)
When processing a batch, if all events fail to encode or the entire batch is too large, everything's fine. However, if specifically the first event in the batch is too large (or JSON fails to encode) but others are encoded just fine, then the batch processing code can panic when it tries to get the API host and other shared values off the first event in the batch (that is now nil).

This change walks the list of events until it finds the first valid event (and there is guaranteed to be a valid event because the number of encoded events is >0) and uses the shared values from that event.

The added test fails with the expected panic without the code change and passes with the change.