feat: Add AddFieldIfUnset method to events
In beeline-go, trace-level fields take precedence over span-level
fields. This is because we add the trace-level fields to the span at
the very end of it's life, just before it's sent.

While we could move the point at which trace-level fields are added to
spans, that feels... like too much backwards incompatibility, and
instead I'd like to propose having trace-level fields added as
"defaults" - so we won't override the fields that are added to the
specific span.